refactor(typescript): simplify structure, add mjs extensions
