+++ packages/eslint-config-typescript/src/index.mjs
@@ -4,7 +4,13 @@
  */
 
 import base from '@vanyauhalin/eslint-config-base';
-import { rules } from './rules';
+import { rules as eslintPluginImport } from './eslint-plugin-import';
+import { rules as typescriptEslint } from './typescript-eslint';
+
+const rules = {
+  ...eslintPluginImport,
+  ...typescriptEslint,
+};
 
 /**
  * Helper function for adding rules.
