--- packages/eslint-config-typescript/src/rules/index.js
@@ -1,11 +1,0 @@
-import { rules as eslintPluginImport } from './eslint-plugin-import';
-import { rules as typescriptEslint } from './typescript-eslint';
-
-const rules = {
-  ...eslintPluginImport,
-  ...typescriptEslint,
-};
-
-export {
-  rules,
-};
