+++ packages/eslint-config-typescript/src/typescript-eslint.mjs
